requirements.txt
streamlit>=1.32.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
langchain-google-genai>=1.0.1
pinecone-client>=3.2.2
langchain-pinecone>=0.1.0
beautifulsoup4>=4.12.3
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.1
tenacity>=8.2.0
